requirements.txt
streamlit
matplotlib
openai
python-dotenv
ast
